+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * chart: fix command handling
 * ng: fix input timeout to also accept int parameters
 * ng: fix `http_input` `collect_meta` leading to shared dicts between events
+* ng: fix error output structure to stay consistent with non-ng
 * generic_adder: allow None as valid input via `add`
 
 ## 20.0.0